Convert Query-level errors flag into max retries integer argument
Before this commit, there was a flag signaling if the Query-level errors should be retried or not.
Now, there's an integer argument representing the max number of Query-level errors retries.

Tenacity now looks at two settings when deciding if it should stop retrying a Query-level error:
- max timeout
- max attempts

This setting is ignored when retries are disabled (`handle_retries=False` or `--disable-retries`).

We've also took the chance to change Domain Occupied stop condition from stop never to the same condition
as other Query-level errors (network timeout).