Reraise last exception in retries instead of RetryError
It makes no sense to raise tenacity.RetryError:
* It exposes an internal about retrials
* It makes more difficult error handling by the client
* It makes the code more complex

Instead, the last exception obtained in the last attempt is reraised.
This way the client will only face with meaningful exceptions
(e.g. RequestError, or other connection errors). The number
of attempts is returned in the stats, so this is not a problem.